requirements.txt
SQLAlchemy==2.0.40
psycopg2==2.9.10
psycopg[binary]
python-dotenv==1.1.0
Flask.Migrate==4.1.0
